% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGST.R
\docType{package}
\name{eGST-package}
\alias{eGST-package}
\alias{_PACKAGE}
\title{eGST (eQTL-based Genetic Sub-Typer): A statistical method to identify tissue-specific
genetic subtypes of complex trait.}
\description{
If multiple tissue or cell-type specific biological pathways underlie a phenotype,
it can be possible to identify the subtype of the phenotype which is defined by heterogeneous
tissue or cell-type -specific genetic architecture of the trait. For example, an individual's
body mass index (BMI)
can be regulated more by the set of genes higher expressed in brain compared to the set
of genes higher expressed in adipose, and vice versa. We propose a method eGST which learns about
such subtype
structure of a complex trait based on individual-level data of marginal phenotype
and genotypes of sets of expression quantitative trait loci (eQTLs) each corresponding
to the set of genes higher expressed in a tissue. The method is based on a finite mixture model
under Bayesian framework and implements the
maximum a posteriori (MAP) expectation-maximization (EM) algorithm for estimation of
subtype posterior probability across individuals and other model parameters.
}
\section{Functions}{

\describe{
\item{\code{\link{eGST}}}{It estimates the posterior probability that an individual's phenotype is a tissue-specific
genetic subtype. The phenotype across individuals can be classified as tissue-specific genetic subtype based on
the estimated subtype posterior probability across individuals.}}
}

\references{
Majumdar A, Giambartolomei C, Cai N, Freund MK, Haldar T, J Flint, Pasaniuc B (2019) Leveraging eQTLs to identify
tissue-specific genetic subtype of complex trait. bioRxiv.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ArunabhaCodes/eGST}
  \item Report bugs at \url{https://github.com/ArunabhaCodes/eGST/issues}
}

}
\author{
\strong{Maintainer}: Arunabha Majumdar \email{statgen.arunabha@gmail.com}

Authors:
\itemize{
  \item Tanushree Haldar \email{tanushree.haldar@gmail.com}
  \item Bogdan Pasaniuc \email{pasaniuc@ucla.edu}
}

}
