% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMAPEM.R
\name{eGST}
\alias{eGST}
\title{Run eGST.}
\usage{
eGST(pheno, geno, tissues, logLimprovement = 10^(-8),
  seed_choice = sample(1:1000, size = 1), nIter = 100)
}
\arguments{
\item{pheno}{A numeric vector of length N where N is the number of individuals. It
contains the GWAS phenotype values of individuals. No default.}

\item{geno}{A list with K elements where K is the number of tissues. Each element of
geno is the genotype matrix of the eQTLs specific to a tissue in the GWAS cohort. So j-th element of geno is N by Mj matrix
containing the genotype data of N individuals (rows) at Mj eQTLs (columns) specific to j-th tissue.
Each eQTL is a bi-allelic SNP with minor allele frequency > 0.01.
Genotypes at each eQTL should be normalized across N individuals.
If 0/1/2 valued genotype matrix is provided, it is internally normalized. No default.}

\item{tissues}{A character vector of length K. It contains the names of tissues of interest in subtyping.
The order of tissues in this vector must match the order of tissues in the previous argument 'geno'. No default.}

\item{logLimprovement}{A positive real number specifying the minimum possible improvement
of data log-likelihood in MAP-EM stopping criterion. Default \eqn{10^(-8)}.}

\item{seed_choice}{An integer providing the choice of random seed for initialization in MAP-EM algorithm. Default is
an integer randomly selected in (1,...,1000).}

\item{nIter}{An integer providing the maximum number of iterations allowed in the MAP-EM algorithm. Default is 100.}
}
\value{
The output produced by \code{\link{eGST}} is a list which consists of various components.
\item{gamma}{A N by K matrix providing the tissue-specific subtype posterior probability of N individuals across K tissues.}
\item{alfa}{Tissue-specific intercepts/means of the trait.}
\item{beta}{Tissue-specific eQTLs' genetic effect on the trait.}
\item{sigma_g}{Square root of variance of tissue-specific per-eQTL genetic effect on the trait.}
\item{sigma_e}{Square root of error variance of tissue-specific subtype of the trait
which remains unexplained by the tissue-specific eQTLs.}
\item{m}{Number of tissue-specific eQTLs.}
\item{logL}{log-likelihood of the data.}
}
\description{
Run eGST to estimate the posterior probability that an individual's phenotype is a tissue-specific genetic subtype.
}
\examples{
data(ExamplePhenoData)
pheno <- ExamplePhenoData
head(pheno)
data(ExampleEQTLgenoData)
geno <- ExampleEQTLgenoData
geno[[1]][1:5,1:5]
geno[[2]][1:5,1:5]
tissues <- paste("tissue", 1:2, sep = "")
result <- eGST(pheno, geno, tissues)
str(result)

}
\references{
Majumdar A, Giambartolomei C, Cai N, Freund MK, Haldar T, J Flint, Pasaniuc B (2019) Leveraging eQTLs to identify
tissue-specific genetic subtype of complex trait. bioRxiv.
}
