% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleEQTLgenoData}
\alias{ExampleEQTLgenoData}
\title{An example of tissue-specific eQTLs genotype data.}
\format{A list of two numeric matrix each having 1000 rows (individuals) and 100 columns (eQTLs):}
\usage{
data(ExampleEQTLgenoData)
}
\description{
ExampleEQTLgenoData is a list with two elements each containing a 1000 by 100 ordered genotype matrix.
Each matrix provides the genotype data of the 1000 individuals at 100 tissue-specific eQTLs for each tissue.
}
\examples{
data(ExampleEQTLgenoData)
geno <- ExampleEQTLgenoData
}
\keyword{datasets}
