% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExamplePhenoData}
\alias{ExamplePhenoData}
\title{An example of phenotype data.}
\format{A numeric vector of length 1000 containing the phenotype data of 1000 individuals.}
\usage{
data(ExamplePhenoData)
}
\description{
ExamplePhenoData is a simulated vector of phenotype for 1000 individuals.
In this simulated example dataset, we have considered two tissues and corresponding
sets of 100 tissue-specific eQTLs each.
First half of 1000 individuals' phenotypes were simulated to have genetic effect from the first tissue
specific eQTLs, but no effect from the second tissue-specific eQTLs. Hence first 500 individuals
were assigned to the first tissue. Similarly, second half of
the 1000 individuals were simulated to have genetic effect from the second-tissue specific eQTLs and hence
they were assigned to the second tissue.
}
\examples{
data(ExamplePhenoData)
pheno <- ExamplePhenoData
}
\keyword{datasets}
